requirements.txt
ansible
google-cloud-secret-manager
google_crc32c